+++ components/deployment-settings/index.tsx
@@ -179,6 +179,7 @@ export function DeploymentSettingsModal({
                 onChange={handleGeneralChange}
                 projectId={projectId}
                 deploymentId={deployment.id}
+                slug={deployment.slug}
                 projects={projects}
                 onProjectChange={setProjectId}
               />